+++ packages/base/src/index.ts
@@ -3,16 +3,15 @@
  */
 import { createComponentStyles } from './lib/createComponentStyles';
 import { CssSizeVariables, CssSizeVariablesNames, cssVariablesStyles } from './lib/CssSizeVariables';
+import * as Device from './lib/Device';
+import * as hooks from './lib/hooks';
 import { LOG_LEVEL, Logger } from './lib/Logger';
+import * as spacing from './lib/spacing';
 import { StyleClassHelper } from './lib/StyleClassHelper';
 import { ThemingParameters } from './lib/ThemingParameters';
 import { useConsolidatedRef } from './lib/useConsolidatedRef';
 import { usePassThroughHtmlProps } from './lib/usePassThroughHtmlProps';
 import { deprecationNotice, enrichEventWithDetails, getScrollBarWidth } from './lib/Utils';
-
-export * as Device from './lib/Device';
-export * as hooks from './lib/hooks';
-export * as spacing from './lib/spacing';
 
 export {
   createComponentStyles,
@@ -27,5 +26,8 @@ export {
   CssSizeVariables,
   cssVariablesStyles,
   ThemingParameters,
-  enrichEventWithDetails
+  enrichEventWithDetails,
+  Device,
+  hooks,
+  spacing
 };