+++ packages/main/src/components/AnalyticalTable/TableBody/VirtualTableBody.tsx
@@ -22,6 +22,8 @@ interface VirtualTableBodyProps {
   overscanCountHorizontal: number;
   renderRowSubComponent: (row?: any) => ReactNode;
   popInRowHeight: number;
+  isRtl: boolean;
+  markNavigatedRow?: (row?: Record<any, any>) => boolean;
 }
 
 export const VirtualTableBody = (props: VirtualTableBodyProps) => {
@@ -41,7 +43,9 @@ export const VirtualTableBody = (props: VirtualTableBodyProps) => {
     parentRef,
     overscanCountHorizontal,
     renderRowSubComponent,
-    popInRowHeight
+    popInRowHeight,
+    markNavigatedRow,
+    isRtl
   } = props;
 
   const rowSubComponentsHeight = useRef({});
@@ -159,7 +163,10 @@ export const VirtualTableBody = (props: VirtualTableBodyProps) => {
   const popInColumn = useMemo(
     () =>
       visibleColumns.filter(
-        (item) => item.id !== '__ui5wcr__internal_highlight_column' && item.id !== '__ui5wcr__internal_selection_column'
+        (item) =>
+          item.id !== '__ui5wcr__internal_highlight_column' &&
+          item.id !== '__ui5wcr__internal_selection_column' &&
+          item.id !== '__ui5wcr__internal_navigation_column'
       )[0],
     [visibleColumns]
   );
@@ -196,6 +203,7 @@ export const VirtualTableBody = (props: VirtualTableBodyProps) => {
         }
         prepareRow(row);
         const rowProps = row.getRowProps();
+        const isNavigatedCell = markNavigatedRow(row);
         const RowSubComponent = typeof renderRowSubComponent === 'function' ? renderRowSubComponent(row) : null;
         return (
           <div
@@ -220,6 +228,12 @@ export const VirtualTableBody = (props: VirtualTableBodyProps) => {
             )}
             {columnVirtualizer.virtualItems.map((virtualColumn, index) => {
               const cell = row.cells[virtualColumn.index];
+              const directionStyles = isRtl
+                ? {
+                    transform: `translateX(-${virtualColumn.start}px)`,
+                    right: 0
+                  }
+                : { transform: `translateX(${virtualColumn.start}px)`, left: 0 };
 
               if (!cell) {
                 return null;
@@ -231,7 +245,8 @@ export const VirtualTableBody = (props: VirtualTableBodyProps) => {
               let contentToRender;
               if (
                 cell.column.id === '__ui5wcr__internal_highlight_column' ||
-                cell.column.id === '__ui5wcr__internal_selection_column'
+                cell.column.id === '__ui5wcr__internal_selection_column' ||
+                cell.column.id === '__ui5wcr__internal_navigation_column'
               ) {
                 contentToRender = 'Cell';
               } else if (isTreeTable || RowSubComponent) {
@@ -252,14 +267,13 @@ export const VirtualTableBody = (props: VirtualTableBodyProps) => {
                     ...cellProps.style,
                     position: 'absolute',
                     width: `${virtualColumn.size}px`,
-                    transform: `translateX(${virtualColumn.start}px)`,
                     top: 0,
-                    left: 0
+                    ...directionStyles
                   }}
                 >
                   {popInRowHeight !== internalRowHeight && popInColumn.id === cell.column.id
                     ? cell.render('PopIn', { contentToRender, internalRowHeight })
-                    : cell.render(contentToRender)}
+                    : cell.render(contentToRender, isNavigatedCell === true ? { isNavigatedCell } : {})}
                 </div>
               );
             })}