+++ packages/main/src/components/AnalyticalTable/index.tsx
@@ -1,5 +1,5 @@
 import { createComponentStyles } from '@ui5/webcomponents-react-base/lib/createComponentStyles';
-import { useIsomorphicLayoutEffect } from '@ui5/webcomponents-react-base/lib/hooks';
+import { useIsomorphicLayoutEffect, useIsRTL } from '@ui5/webcomponents-react-base/lib/hooks';
 import { StyleClassHelper } from '@ui5/webcomponents-react-base/lib/StyleClassHelper';
 import { ThemingParameters } from '@ui5/webcomponents-react-base/lib/ThemingParameters';
 import { usePassThroughHtmlProps } from '@ui5/webcomponents-react-base/lib/usePassThroughHtmlProps';
@@ -49,6 +49,7 @@ import { useDragAndDrop } from './hooks/useDragAndDrop';
 import { useDynamicColumnWidths } from './hooks/useDynamicColumnWidths';
 import { usePopIn } from './hooks/usePopIn';
 import { useRowHighlight } from './hooks/useRowHighlight';
+import { useRowNavigationIndicators } from './hooks/useRowNavigationIndicator';
 import { useRowSelectionColumn } from './hooks/useRowSelectionColumn';
 import { useSingleRowStateSelection } from './hooks/useSingleRowStateSelection';
 import { useStyling } from './hooks/useStyling';
@@ -130,7 +131,11 @@ export interface TableProps extends Omit<CommonProps, 'title'> {
    */
   alternateRowColor?: boolean;
   /**
-   * Flag whether the table should add an extra column for displaying row highlights, based on the `highlightField` prop.
+   * Flag whether the table should add an extra column at the end of the rows for displaying a navigation highlight.
+   */
+  withNavigationHighlight?: boolean;
+  /**
+   * Flag whether the table should add an extra column at the start of the rows for displaying row highlights, based on the `highlightField` prop.
    */
   withRowHighlight?: boolean;
   /**
@@ -203,6 +208,13 @@ export interface TableProps extends Omit<CommonProps, 'title'> {
   infiniteScrollThreshold?: number;
 
   // events
+
+  /**
+   * This callback can be used to programmatically show an indicator for navigated rows. It has no effect if `withNavigationHighlight` is not set.
+   *
+   * __Must be memoized!__
+   */
+  markNavigatedRow?: (row?: Record<any, any>) => boolean;
   /**
    * Fired when the sorting of the rows changes.
    */
@@ -214,11 +226,15 @@ export interface TableProps extends Omit<CommonProps, 'title'> {
   /**
    * Fired when a row is selected or unselected.
    */
-  onRowSelected?: (e?: CustomEvent<{ allRowsSelected?: boolean; row?: unknown; isSelected?: boolean }>) => any;
+  onRowSelected?: (e?: CustomEvent<{ allRowsSelected?: boolean; row?: unknown; isSelected?: boolean }>) => void;
+  /**
+   * Fired when a row is clicked
+   */
+  onRowClick?: (e?: CustomEvent<{ row?: unknown }>) => void;
   /**
    * Fired when a row is expanded or collapsed
    */
-  onRowExpandChange?: (e?: CustomEvent<{ row: unknown; column: unknown }>) => any;
+  onRowExpandChange?: (e?: CustomEvent<{ row: unknown; column: unknown }>) => void;
   /**
    * Fired when the columns order is changed.
    */
@@ -290,6 +306,7 @@ const AnalyticalTable: FC<TableProps> = forwardRef((props: TableProps, ref: Ref<
     selectionMode,
     selectionBehavior,
     onRowSelected,
+    onRowClick,
     onSort,
     reactTableOptions,
     tableHooks,
@@ -311,6 +328,8 @@ const AnalyticalTable: FC<TableProps> = forwardRef((props: TableProps, ref: Ref<
     scaleWidthMode,
     withRowHighlight,
     highlightField,
+    withNavigationHighlight,
+    markNavigatedRow,
     groupable,
     sortable,
     filterable,
@@ -326,6 +345,8 @@ const AnalyticalTable: FC<TableProps> = forwardRef((props: TableProps, ref: Ref<
 
   const [analyticalTableRef, reactWindowRef] = useTableScrollHandles(ref);
   const tableRef: RefObject<DivWithCustomScrollProp> = useRef();
+
+  const isRtl = useIsRTL(analyticalTableRef);
 
   const getSubRows = useCallback((row) => row[subRowsKey] || [], [subRowsKey]);
 
@@ -370,6 +391,7 @@ const AnalyticalTable: FC<TableProps> = forwardRef((props: TableProps, ref: Ref<
         selectionBehavior,
         classes,
         onRowSelected,
+        onRowClick,
         onRowExpandChange,
         isTreeTable,
         alternateRowColor,
@@ -377,6 +399,8 @@ const AnalyticalTable: FC<TableProps> = forwardRef((props: TableProps, ref: Ref<
         loading,
         withRowHighlight,
         highlightField,
+        withNavigationHighlight,
+        markNavigatedRow,
         renderRowSubComponent
       },
       ...reactTableOptions
@@ -391,6 +415,7 @@ const AnalyticalTable: FC<TableProps> = forwardRef((props: TableProps, ref: Ref<
     useRowSelectionColumn,
     useSingleRowStateSelection,
     useRowHighlight,
+    useRowNavigationIndicators,
     useDynamicColumnWidths,
     useStyling,
     useToggleRowExpand,
@@ -448,6 +473,10 @@ const AnalyticalTable: FC<TableProps> = forwardRef((props: TableProps, ref: Ref<
       parentHeightObserver.disconnect();
     };
   }, [updateTableClientWidth, updateRowsCount]);
+
+  useIsomorphicLayoutEffect(() => {
+    dispatch({ type: 'IS_RTL', payload: { isRtl } });
+  }, [isRtl]);
 
   useIsomorphicLayoutEffect(() => {
     updateTableClientWidth();
@@ -546,6 +575,7 @@ const AnalyticalTable: FC<TableProps> = forwardRef((props: TableProps, ref: Ref<
 
   const [dragOver, handleDragEnter, handleDragStart, handleDragOver, handleOnDrop, handleOnDragEnd] = useDragAndDrop(
     props,
+    isRtl,
     setColumnOrder,
     tableState.columnOrder,
     tableState.columnResizing,
@@ -556,6 +586,7 @@ const AnalyticalTable: FC<TableProps> = forwardRef((props: TableProps, ref: Ref<
     'onSort',
     'onGroup',
     'onRowSelected',
+    'onRowClick',
     'onRowExpandChange',
     'onColumnsReordered',
     'onLoadMore'
@@ -653,6 +684,7 @@ const AnalyticalTable: FC<TableProps> = forwardRef((props: TableProps, ref: Ref<
                   onDragEnter={handleDragEnter}
                   onDragEnd={handleOnDragEnd}
                   dragOver={dragOver}
+                  isRtl={isRtl}
                 />
               )
             );
@@ -660,6 +692,7 @@ const AnalyticalTable: FC<TableProps> = forwardRef((props: TableProps, ref: Ref<
           {loading && props.data?.length > 0 && <LoadingComponent style={{ width: `${totalColumnsWidth}px` }} />}
           {loading && props.data?.length === 0 && (
             <TablePlaceholder
+              isRtl={isRtl}
               columns={tableInternalColumns.filter(
                 (col) => (col.isVisible ?? true) && !tableState.hiddenColumns.includes(col.accessor)
               )}
@@ -705,6 +738,8 @@ const AnalyticalTable: FC<TableProps> = forwardRef((props: TableProps, ref: Ref<
                 visibleColumnsWidth={visibleColumnsWidth}
                 overscanCountHorizontal={overscanCountHorizontal}
                 renderRowSubComponent={renderRowSubComponent}
+                markNavigatedRow={markNavigatedRow}
+                isRtl={isRtl}
               />
             </VirtualTableBodyContainer>
           )}
@@ -758,6 +793,7 @@ AnalyticalTable.defaultProps = {
   visibleRows: 15,
   subRowsKey: 'subRows',
   highlightField: 'status',
+  markNavigatedRow: () => false,
   selectedRowIds: {},
   onGroup: () => {},
   onRowExpandChange: () => {},