+++ packages/main/src/components/AnalyticalTable/tableReducer/stateReducer.ts
@@ -1,5 +1,32 @@
+import { actions } from 'react-table';
+
 export const stateReducer = (newState, action) => {
   const { payload } = action;
+
+  if (newState.isRtl && action.type === actions.columnResizing) {
+    const { clientX } = action;
+    const { startX, columnWidth, headerIdWidths } = newState.columnResizing;
+
+    const deltaX = startX - clientX;
+    const percentageDeltaX = deltaX / columnWidth;
+
+    const newColumnWidths = {};
+
+    headerIdWidths.forEach(([headerId, headerWidth]) => {
+      newColumnWidths[headerId] = Math.max(headerWidth + headerWidth * percentageDeltaX, 0);
+    });
+
+    return {
+      ...newState,
+      columnResizing: {
+        ...newState.columnResizing,
+        columnWidths: {
+          ...newState.columnResizing.columnWidths,
+          ...newColumnWidths
+        }
+      }
+    };
+  }
 
   switch (action.type) {
     case 'TABLE_RESIZE':
@@ -14,6 +41,8 @@ export const stateReducer = (newState, action) => {
       return { ...newState, popInColumns: payload };
     case 'INTERACTIVE_ROWS_HAVE_POPIN':
       return { ...newState, interactiveRowsHavePopIn: payload };
+    case 'IS_RTL':
+      return { ...newState, isRtl: payload.isRtl };
     default:
       return newState;
   }